Share prefab source normalization
