+++ Sources/OvEditor/src/OvEditor/Core/EditorActions.cpp
@@ -59,40 +59,6 @@ namespace
 {
 	constexpr std::string_view kDefaultMaterialPath = ":Materials\\Default.ovmat";
 
-	void NormalizePrefabSourcesRecursively(
-		OvCore::ECS::Actor& p_actor,
-		const std::string& p_inheritedPrefabSource)
-	{
-		if (p_actor.HasPrefabSource())
-		{
-			const std::string currentPrefabSource = p_actor.GetPrefabSource();
-			if (currentPrefabSource == p_inheritedPrefabSource)
-			{
-				p_actor.SetPrefabSource("?");
-			}
-		}
-
-		const std::string nextInheritedPrefabSource =
-			p_actor.HasPrefabSource() ? p_actor.GetPrefabSource() : p_inheritedPrefabSource;
-
-		for (auto* child : p_actor.GetChildren())
-		{
-			NormalizePrefabSourcesRecursively(*child, nextInheritedPrefabSource);
-		}
-	}
-
-	void SetRootPrefabSourceAndNormalizeChildren(
-		OvCore::ECS::Actor& p_rootActor,
-		const std::string& p_newRootPrefabSource)
-	{
-		p_rootActor.SetPrefabSource(p_newRootPrefabSource);
-
-		for (auto* child : p_rootActor.GetChildren())
-		{
-			NormalizePrefabSourcesRecursively(*child, p_newRootPrefabSource);
-		}
-	}
-
 	OvCore::ECS::Actor* ResolvePrefabInstanceRoot(OvCore::ECS::Actor& p_actor)
 	{
 		auto* resolvedRoot = &p_actor;
@@ -1189,7 +1155,7 @@ void OvEditor::Core::EditorActions::SaveActorAsPrefab(OvCore::ECS::Actor& p_acto
 	}
 
 	const std::string prefabSourcePath = GetResourcePath(p_path);
-	SetRootPrefabSourceAndNormalizeChildren(p_actor, prefabSourcePath);
+	OvCore::SceneSystem::PrefabOperations::SetRootPrefabSourceAndNormalizeChildren(p_actor, prefabSourcePath);
 
 	OVLOG_INFO("Prefab saved to: " + p_path);
 }
@@ -1212,7 +1178,7 @@ OvCore::ECS::Actor* OvEditor::Core::EditorActions::InstantiatePrefab(const std::
 
 	if (instantiatedRoot)
 	{
-		SetRootPrefabSourceAndNormalizeChildren(*instantiatedRoot, prefabSourcePath);
+		OvCore::SceneSystem::PrefabOperations::SetRootPrefabSourceAndNormalizeChildren(*instantiatedRoot, prefabSourcePath);
 
 		const std::string prefabName = realPath.stem().string();
 		if (!prefabName.empty())
